% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_annot_lanes.R
\name{wb_annot_lanes}
\alias{wb_annot_lanes}
\title{Add a header annotating Western blot lanes}
\usage{
wb_annot_lanes(img, annot, n_lanes = 8, text_size = 20)
}
\arguments{
\item{img}{A \code{magick-image}}

\item{annot}{A \code{data.frame}, where each \emph{row} represents a lane and each
\emph{column} represents a variable. The first row will annotate the first lane
(left to right)}

\item{n_lanes}{Numeric. The number of lanes in the image}

\item{text_size}{Size of text. Will be automatically downscalled if too big
to fit.}
}
\value{
A \code{magick-image}
}
\description{
Add a header annotating Western blot lanes
}
