% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_list.R
\name{wb_list}
\alias{wb_list}
\title{List files in a directory}
\usage{
wb_list(
  path,
  recursive = FALSE,
  user = Sys.getenv("WEBBER_USR"),
  wb_path = Sys.getenv("WEBBER_WBPATH")
)
}
\arguments{
\item{path}{Character. The path to the image, sans \code{wb_path}}

\item{recursive}{Logical. Should folders of folders be recursively indexed into?}

\item{user}{Whose folder should this file be pulled from?}

\item{wb_path}{Character. Path to the top level directory that contains all western blots.}
}
\value{
a \code{data.frame} with the path to the item on SharePoint
}
\description{
List files in a directory
}
\examples{
\dontrun{
wb_list(
  "2022-02-04_mcconkey_fgfr3",
  recursive = TRUE,
  user = "aragaki-kai",
  wb_path = "Raw Data/ChemiDoc"
)
wb_list(
  "2022-02-04_mcconkey_fgfr3",
  recursive = FALSE,
  user = "aragaki-kai",
  wb_path = "Raw Data/ChemiDoc"
)
}
}
