% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_put.R
\name{wb_put}
\alias{wb_put}
\title{Upload a Western Blot}
\usage{
wb_put(
  path,
  dest_path,
  wb_edit_path = Sys.getenv("WEBBER_WBEDITPATH"),
  user = Sys.getenv("WEBBER_USR"),
  drive = NULL
)
}
\arguments{
\item{path}{Character. Path to the local file.}

\item{dest_path}{Character. Path to upload the file. See details.}

\item{wb_edit_path}{Character. Path where Western blot edits are kept.}

\item{user}{Character. User directory name.}

\item{drive}{An optional \code{ms_drive} object. If none supplied, will make one
automatically.}
}
\value{
Character. Path to the file on Sharepoint.
}
\description{
This is a thin wrapper around the \code{upload_file} function from
\code{Microsoft365R}, it just has defaults and safeguards tailored to uploading
Western Blots. You could theoretically upload anything this way. It WILL
overwrite without prompting.
}
\details{
The uploaded file location is produced via \code{fs::path(wb_edit_path, user, dest_path)}
}
\examples{
\dontrun{
wb <- wb_get(
"2022-06-22/h3.tif",
tempdir(),
user = "aragaki-kai",
wb_path = "Raw Data/licor"
)

wb |>
}
}
