% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_get.R
\name{wb_get}
\alias{wb_get}
\title{Download a Western blot by path}
\usage{
wb_get(
  path,
  dest_dir,
  overwrite = FALSE,
  user = Sys.getenv("WEBBER_USR"),
  wb_path = Sys.getenv("WEBBER_WBPATH"),
  quiet = FALSE
)
}
\arguments{
\item{path}{Character. The path to the image/directory, sans \code{wb_path}.}

\item{dest_dir}{Character. The path to save the image/directory. If it
doesn't have an extension, it's assumed you mean a folder, and that folder
already exists.}

\item{overwrite}{Logical. Should files be overwritten if they already exist?}

\item{user}{Character. Whose folder this file should be pulled from.}

\item{wb_path}{Character. Path to the top level directory that contains all
western blots.}

\item{quiet}{Logical. Should this function not message upon finding existing
files?}
}
\value{
Character. The local path to the file or directory.
}
\description{
Downloads a Western blot image and returns its path. If it already exists in
the specified location, no downloading will occur unless \code{overwrite = TRUE}.
Will return the local path to the file.
}
\details{
This is a thin wrapper around \code{bladdr::get_gbci_file}, and has hard-coded
paths - so it's more a quality-of-life function.
}
\examples{
\dontrun{
wb_get(
  "2022-06-22/h3.tif",
  tempdir(),
  user = "aragaki-kai",
  wb_path = "Raw Data/licor"
)
}
}
